refactor(middleware): add getPaths helper as separate util
